# Conflicts:
#	.gitlab-ci.yml
#	src/aarith/integer/integer_operations.hpp
#	src/aarith/integer/integer_ranges.hpp
#	src/aarith/integer/integers.hpp
#	tests/integer/integer-comparisons-test.cpp

+++ src/aarith/core/number_utils.hpp
@@ -94,7 +94,7 @@ constexpr size_t pow(const size_t base, const size_t exponent)
     const size_t first_bit = first_set_bit(n);
     if (first_bit > 0)
     {
-        return 1UL << (first_bit - 1);
+        return static_cast<size_t>(1) << (first_bit - 1);
     }
     else
     {
